# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "List articles",
                "operationId": "658936351861b1e6be24348b68cfd902",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Filter by author",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "published_at",
                        "in": "query",
                        "description": "Filter by date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "title": "Sample",
                                            "content": "Text",
                                            "author": "Author",
                                            "published_at": "2024-01-01T00:00:00Z",
                                            "created_at": "2024-01-01T00:00:00Z",
                                            "updated_at": "2024-01-01T00:00:00Z"
                                        }
                                    ],
                                    "meta": [],
                                    "links": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article details",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                },
                                "example": {
                                    "id": 1,
                                    "title": "Sample",
                                    "content": "Text",
                                    "author": "Author",
                                    "published_at": "2024-01-01T00:00:00Z",
                                    "created_at": "2024-01-01T00:00:00Z",
                                    "updated_at": "2024-01-01T00:00:00Z"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/feed/personalized": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get personalized article feed based on user preferences",
                "operationId": "fc1376b0cc7f134473a1e26cbfe83e45",
                "responses": {
                    "200": {
                        "description": "Personalized article feed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "title": "Sample",
                                            "content": "Text",
                                            "author": "Author",
                                            "published_at": "2024-01-01T00:00:00Z",
                                            "created_at": "2024-01-01T00:00:00Z",
                                            "updated_at": "2024-01-01T00:00:00Z"
                                        }
                                    ],
                                    "meta": [],
                                    "links": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "required": [
                    "title",
                    "content",
                    "author"
                ],
                "properties": {
                    "id": {
                        "description": "Article ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the article",
                        "type": "string"
                    },
                    "content": {
                        "description": "Content of the article",
                        "type": "string"
                    },
                    "author": {
                        "description": "Author of the article",
                        "type": "string"
                    },
                    "published_at": {
                        "description": "Publication date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}